requirements.txt
python-socketio==5.0.1
aiortc==1.12.0
structlog==25.4.0
py-cord[voice]